Added gridCVA function
Added a new function that allow for a grid/sliding-window approach to
computing lattice rotation tensors for a *SINGLE* phase. The function
requires a set of EBSD data of a single phase and an integer to specify
the window “padding” used for the sliding window. For example, if pad =
1, then the rotation tensor will be calculated based on 9 orientations
– the central (spatially) point and the 8 immediately adjacent
solutions. If there are non-indexed solutions, the tensor will still be
calculated, unless the window contains fewer than 3 solutions (3+
solutions are required for the PGR calculation). The function will
return an EBSD dataset that matches the points for which tensors were
created. Furthermore, the returned ebsd variable is appended with the
results so that they can be easily queried and plotted. This function
is likely to be updated again in the near future, which may result in
slight changes to the function call and syntax. However, the core
functionality is not likely to change, and I am sharing it now for
colleagues who have expressed interest.