Added a filtered version (& a Note about appended grainset)
Included a "calcGrainsDispersionFilter" script as an example of one way to further explore intragranular dispersion patterns through weighting or filtering the results. In this *example* script, the applied filter excludes any results for which the ratio between the primary and seconday principal axes of the rotation ellipsoid/tensor is not greater than 1.4. The value 1.4 is completely arbitrary, and I selected it only as an example using a value familiar to me as a threshold for shape preferred orientation of grain shapes (e.g., only considering ellipse axes from grain ellipses that have a ratio > 1.4).

Note:
Recently I have updated the repository to include a new function that allows for a new method/syntax for retrieving the results. In this version, the result is a new grainset variable of the grains that were analyzed, and the grainset is appended with CVA data as "dot-accessible" properties. For example, to access the CVA axis for a given grain in the new grainset, you could type: gCVA.CVA
I prefer this approach, and I believe it will make working with the data easier for most users. Additional note: This new version excludes any grains tht do not have more than 2  orientation/solutions from the ebsd dataset. It also excludes any grains with a grain orientation spread (GOS) that is not greater than 0 degrees. These two things are necessary for computing the best-fit rotation tensor.