IMPORTANT FIX: Accounts for change in mtex versions 5.2+
Some users noted that CVA results were different for some datasets (phase specific I think) if run in mtex 5.2+. It seems the culprit is a subtle -- but crucial -- difference in how 5.2+ treats orientations. the necessary change was to now extract the "rotations" of the orientations, to avoid potential confounding symmetry issues.
Please update!!