Overhaul of the grid/kernel-based CVA function and modification to grain-scale function
Kernel/grid function ('gridCVA.m') adds new functionality and now operates with a 3x3 sliding-window (as opposed to the segmentation approach in the previous version/s). This change makes the results much more usable at the expense of computational time (it is quite slow).  The function also now works with all phases at once (rather than need to isolate a single phase), and it allows an optional input of a related grainset, which will result in the removal of pixels adjacent to grain boundaries from the analysis and results. The kernel-scale function also adds additional kernel-scale metrics that are assigned/appended to the center-pixel of the kernel in the resultant EBSD variable. These include: mean orientation, orientation spread, mean axis of misorientation of pixels to mean orientation.

Both the grain-scale function and the kernel/grid function add inverted CVA results that are returned as Miller directions in the crystal reference frame.