Added inverseCVA function
invCVA.m handles the inversion of CVA results for users from an input variable that contains results from the CVA functions grainsCVA.m and gridCVA.m. Note, to work with gridCVA results requires the slightly modified version of gridCVA included in this update.
The invCVA function includes a section of code for plotting the populations of the three different eigenvectors of the dispersion tensor. The first eigenvector is the dominant axis of dispersion/rotation (the CVA axis). The second and third eigenvectors are the axes of intermediate and least dispersion, respectively. These may be useful for constraining the crystal-scale kinematics, beyond just the CVA axes.