Create comboPlotPGA.m
Add function to plot all the eigenvectors from the 2nd rank dispersion tensors. For context, each CVA is the maximum eigenvector of the dispersion tensor... but there are 2 other potentially relevant eigenvectors that are ignored by vanilla "CVA analysis". Those vectors and the computed dispersion tensors from which they are derived are all returned in the grainsCVA() function as properites in the grainset... *THIS* plotting function uses those results to visualize all of the eigenvectors from the dispersion tensors together.