Update to PGA.m to be compatible with both MTEX 5.10.x and 5.11.2
The release of MTEX 5.11.2 broke the PGA function due to syntax changes by the developers that did not maintain compatibility with previous versions. This new function includes a check for the MTEX version being used, and selects the appropriate syntax for proper computation.

NOTE: due to other changes, the gridCVA.m function does not currently work with MTEX 5.11.2. This will hopefully be fixed soon.